# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: MyAgent API
  description: |
    AI Agent 管理平台后端 API。

    ## 认证方式

    除健康检查和认证端点外，所有 API 请求需在 `Authorization` 头携带 Bearer token：

    ```
    Authorization: Bearer <access_token>
    ```

    Token 通过 `/api/auth/login` 或 `/api/auth/register` 获取。

    ## 通用响应格式

    所有标准 API 响应遵循统一信封格式：
    - 成功：`{ "success": true, "data": T, "meta?": { total, page, limit } }`
    - 失败：`{ "success": false, "error": string, "message": string }`

    错误 HTTP 状态码对照表：

    | 状态码 | 含义 | 前端行为 |
    |--------|------|---------|
    | 401 | 未认证 / Token 过期 | 拦截器自动尝试 refresh token |
    | 403 | 无权限 | 跳转 403 页面 |
    | 404 | 资源不存在 | 显示 404 |
    | 409 | 资源冲突 | 提示用户 |
    | 422 | 参数校验失败 | 显示字段级错误 |
    | 500 | 服务器内部错误 | 显示通用错误页 |
  version: "0.2.0"
  contact:
    name: MyAgent Team
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: http://localhost:3000
    description: 前端 Next.js 开发服务器（通过代理转发）
paths:
...
  /api/auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      description: 注册新用户，成功后返回 JWT token 对和用户信息（自动登录）。
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 注册成功，返回 token 和用户信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: 邮箱已被注册
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error409"
        "422":
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error422"
  /api/auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      description: 使用邮箱和密码登录，返回 JWT token 对和用户信息。
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: 邮箱或密码错误 / 账户被禁用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error401"
...
  /api/auth/refresh:
    post:
      tags: [认证]
      summary: 刷新 Token
      description: 使用 refresh_token 获取新的 access_token 和 refresh_token。
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
      responses:
        "200":
          description: Token 刷新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "401":
          description: refresh_token 无效或已过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error401"
  /api/auth/profile:
    get:
      tags: [认证]
      summary: 获取个人信息
      description: 获取当前登录用户的详细信息。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 用户信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/auth/password:
    put:
      tags: [认证]
      summary: 修改密码
      description: 修改当前登录用户的密码。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        "200":
          description: 密码修改成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: 原密码错误或新密码不符合要求
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  # ============================================================================
  # Health
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token，通过登录或注册获取
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 操作成功
...
    Error400:
      type: object
      properties:
        detail:
          type: string
          example: 原密码错误
    Error401:
      type: object
      properties:
        detail:
          type: string
          example: 邮箱或密码错误
...
    Error409:
      type: object
      properties:
        detail:
          type: string
          example: 该邮箱已被注册
...
    Error422:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                example: ["body", "email"]
              msg:
                type: string
                example: Value error, 邮箱格式不正确
              type:
                type: string
                example: value_error
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          description: 用户名
          example: 张三
        email:
          type: string
          format: email
          description: 邮箱
          example: zhangsan@example.com
        password:
          type: string
          minLength: 6
          description: 密码
          example: password123
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: zhangsan@example.com
        password:
          type: string
          example: password123
    TokenRefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: 刷新令牌
    TokenRefreshResponse:
      type: object
      properties:
        access_token:
          type: string
          description: 新的 access_token
        refresh_token:
          type: string
          description: 新的 refresh_token
    PasswordChangeRequest:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          description: 原密码
        new_password:
          type: string
          minLength: 6
          description: 新密码
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: 张三
        email:
          type: string
          example: zhangsan@example.com
        avatar:
          type: string
          default: ""
          example: https://example.com/avatar.png
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access_token（有效期 24 小时）
          example: eyJhbGciOiJIUzI1NiIs...
        refresh_token:
          type: string
          description: JWT refresh_token（有效期 7 天）
          example: eyJhbGciOiJIUzI1NiIs...
        user:
          $ref: "#/components/schemas/UserResponse"

    # ── Chat ────────────────────────────────────────────────────
...
  responses:
    UnauthorizedError:
      description: 未认证或 Token 过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error401"
...